Fixing MacOS building

+++ markup_backend/include/markup_backend/markup.h
@@ -8,7 +8,7 @@
 #include "markup_backend/video.h"
 #include "markup_backend/utils.h"
 
-#include <boost/filesystem.hpp>
+// #include <boost/filesystem.hpp>
 // using namespace boost::filesystem;
 
 struct PipelineRunParams {
@@ -46,7 +46,7 @@ public:
     bool delete_track(size_t id);
 
     size_t add_detection(const Detection& det);
-    size_t add_detection(int frame, int xmin, int ymin, int xmax, int ymax);
+    size_t add_detection(size_t frame, int xmin, int ymin, int xmax, int ymax);
 
     bool delete_detection(size_t track_id, size_t frame_idx);
